+++ libuvxx/include/fs/file.hpp
@@ -37,7 +37,7 @@ namespace uvxx { namespace fs
 
         int64_t file_position_get() const;
 
-        void file_position_set(size_t position) const;
+        void file_position_set(int64_t position) const;
         
         static uvxx::pplx::task<void> delete_async(std::string const& file_name);
